Settable slot object instance for object scanners
Adds a private _slotObjectPtr pointer and setSlotObject() setter method
to enable scanner client to provide own instance of GC_SlotObject
subclass to receive found referent slots in getNextSlot().

This is a much simpler way to implement and use specialized slot
object containers that extend existing GC_SlotObject than that
described in #1638.

In testing so far (and in production testing) the default slot object
is GC_SlotObject::_slotObject and no changes to other code to override
this using new GC_ObjectScanner::setSlotObject() method are included
with this commit.

Downstream work relating to this will be undertaken in evacuator 
development stream and then possibly migrated to other contexts.

Closes: #1638

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>