Select hot field for first slot in object scanner
The aliasing technique employed by scavenger and evacuator has a higher
likelihood of copying the first slot returned by parent object scanner
close to the parent. Likelihood of close copying then tends to drop as 
subsequent child slots are copied. The JIT has a mechanism to identify
hot fields, but this capability is used at cross-purposes in scavenger
and not at all in present evacuator implementation. This change forces
object scanners to preselect a hot field as the first slot to be 
returned in the first call to getNextSlot() after the object scanner is
instantiated. 

This change requires introduction of a new pointer-valued field
_hotScanPtr in GC_ObjectScanner. This will have a greater impact on
system memory footprint for scavenger (which includes an object scanner
in every work packet) than on evacuator (32 object scanner instances per
gc thread).

To test this, I ran 2x32 trials of evacuator with (hot) and without
(cold) this change. The results presented in the table below suggest
that using hot field selection for first scanned slot results in a small
but measurable improvement in benchmark scores, as intended. 

		cold		hot			t-score (2-tailed)
stdev	 1357.72	 1575.42	0.21
mean	30707.15	31169.64	
sum		 1013336	 1028598	

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>