Evacuator: Enable >64 GC threads
This commit enables evacuator controller to manage >64 concurrent
evacuator threads in gencon. Evacuator is disabled by default in OpenJ9
runtimes but may be enabled (-Xgc:recursiveScanOrdering) for performance
testing.

Tested with 5xjbb2005 on 8-core linux/x86-64 fyre vm. Scores look ok:

```
29465
34571
32550
30250
29374
```

Average is 31242 which compares well with averages for 32x runs with
cold and hot evacuator SDKs in #2663 (30707, 31169 -- these were run on
the same platform as above, SDKs were built from this evacuator branch
with all but this commit).

Also tested on a 120-core linux/x86-64 platform with 120 gc threads 
without failure. Tuning for evacuator scalability is a work in progress.

There are a few additional riders with this commit:

- remove redundant return values in stack operators (push/pop/flush)
- remove redundant remainder parameter from getInsideCopyspace()
- remove (assert) redundant NULL checks in findWork()
- ensure more even sizes for split array segments
- add pullWhitespace() method to MM_EvacuatorScanspace

The last rider is forward-looking, pullWhitespace() is not otherwise
referenced in this commit. It will be used in next commit which will
enable interleaving of >1 survivor strata (ages) on the stack. This is
the basis for evacuator support for balanced gc policy. Next commit 
will enable gencon survivor and tenure frames to be interleaved. This 
will provide an opportunity to test interleaving in its simplest
configuration (2 survivor strata) before attempting integration with
balanced policy (24 survivor strata).

Evacuator: Enable >64 GC threads (fix work packet sizes)

At high thread count (120) with low workload (SPECjbb2005) evacuator
performance regressed considerably vs scavenger. Scavenger hierarchical
scanning has a similar problem (high stalled thread counts drive cache
size down to minimum 8k and force breadthfirst copying for all running
threads). The problem is worse for evacuator because it has a lower
minimum work packet size (4k) and releases a lot of small (10s or 100s
of bytes) work packets. Small work packet release is due to a change
included in the original commit that forced running threads to release
an outside copyspace as distributable work in the presence of stalled
evacuators. 

This commit addresses proliferation of small work packets in 4 ways:

1- increase minimum work packet size from 4k to 16k (max is still 64k)
2- more robust dynamic sizing of whitespace and work packets
3- load smallest outside copyspace onto stack before pulling work
4- write large objects <4k to stack if they do not fit in copyspace

The 3rd change forces evacuators with empty worklists to load work from
the outside copyspace with least amount of scan work onto stack and scan
it to completion. This will tend to add work to the other outside
copyspace so it may grow to minimum work packet size and be pulled onto
the worklist where it can be shared with other evacuators. 

The 4th change reduces the number of object-sized whitespace allocations
that occur when the copyspace has <768 bytes free but the object is at
least 768 bytes in size. Evacuator, like scavenger, would have made an
allocation specifically for this object and released the containing
copyspace contents as a work packet. If the object is smaller than a
split array segment, evacuator will now try to copy it into the current
stack frame. This will reduce contention for allocation lock and inhibit
release of small work packets. 

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>